+++ src/CR-Tools/implement/Pypeline/modules/grid.py
@@ -5,7 +5,9 @@
 
 from datetime import datetime
 import pycrtools as cr
-from pytmf import deg2rad, rad2deg, gregorian2jd
+print "Patching grid.py: from pytmf import deg2rad, rad2deg, gregorian2jd"
+#from pytmf import deg2rad, rad2deg, gregorian2jd
+from pytmf import deg2rad, rad2deg
 
 def casaRefcodes(**kwargs):
     """Get CASA reference code and projection from standard WCS parameter