+++ src/CR-Tools/implement/Pypeline/modules/tasks/beamformer.py
@@ -205,6 +205,10 @@ class BeamFormer(tasks.Task):
         "doplot":{default:False,
                   doc:"Plot current spectrum while processing."},
 
+        "newfigure":p_(True,"Create a new figure for plotting for each new instance of the task."),
+
+        "figure":p_(None,"The matplotlib figure containing the plot",output=True),
+
         "plotspec":{default:True,
                     doc:"If **True** plot the beamformed average spectrum at the end, otherwise the time series."},
 
@@ -546,6 +550,8 @@ class BeamFormer(tasks.Task):
         
         if self.doplot:
             plt.ioff()
+            if self.newfigure and not self.figure:
+                self.figure=plt.figure()
 
         original_file_start_number=self.file_start_number
         self.beams.fill(0)